Improve process streams
This makes a number of pretty significant and breaking changes to
process streams, in the hopes of making them more useful, more
consistent and easier to understand.

While we have found a pretty intuitive API for working with the
processes themselves, their stdout and stderr streams have been a bit
neglected. We are already foreseeing some things that we want to do in
bigtest that are difficult to do with the current API.

First of all we create a solid foundation by making the std streams over
Buffers. This is what node exposes by default, and since we would pay a
performance penalty in cases where we unncessarily convert to strings,
we should just expose this.

However, working with raw buffers is tedious, so we expose two smart
mapping functions on iostreams `text()` and `lines()`. `text()` just
converts the buffers into strings. `lines` is a bit fancier and creates
a new stream which emits once for each printed line (waiting until the
line is complete).

Additionally we improve the `buffer` function on `Stream` by making it
return a resource, and also taking an optional limit, so we can create a
bounded buffer using a RingBuffer behind the scenes.

In bigtest we have previously wanted to cache the previous X number of
lines. With this change, something like the following becomes possible,
which was previously pretty challenging:

```
let app = yield exec(options.appCommand);
let lines = yield app.stdout.lines.buffer(1000);

yield lines.filter((l) => l.includes('ready')).expect();

```