✨Add global "onetime" setup to test adapter
We need a way to do global setup before each suite of tests that will
only run once. This corresponds to `beforeAll()` behavior in BDD.

Instead of manually orchestrating scopes via `createScope()` this uses
a tree of spawned tasks: one per test adapter. Whenever a test adapter
runs a test, it will initialize from the top down. A task for the root
adapter will be spawned and initialized with any one-time setup. Then,
its child adapter's scope will be spawned within it, and its one-time
setup run, and so on down to the current adapter.

Once the scope has been initialized (or retrieved if it has already
been initialized) then the test can run by evaluating all of its setup
and body. This maintains a clean hierarchy of scopes that can share
hierarchical tests.

Finally, the adapter destroy method just halts the task associated
with the scope.