🐛 run more than one test
The introduction of one-time setup also introduced a crippling bug
that prevented more than one test from being run. This was because the
actual Efection scope in which each test was being run was shut down
after that test was run. because `runTest()` was a fully encapsulated
task and all scopes were being created as children of the `runTest()`
scope, which by the rules of structured concurrency were killed off
before `runTest()` was complete.

This change makes sure

This is important because it means you can run setup code inside the
test scope and any changes that it makes will be both persistent and
also visible to child scopes of subsequent runs.